% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pred_graph.R
\name{find_nearest_four}
\alias{find_nearest_four}
\title{Find the four nearest neighbours of the spatial field graph}
\usage{
find_nearest_four(pred_coordinates, nn_graph)
}
\arguments{
\item{pred_coordinates}{An sf object with point geometries}

\item{nn_graph}{The output of make_nn_graph}
}
\value{
A list of matrices giving the neighbour indices for each prediction location
}
\description{
Find the four nearest neighbours of the spatial field graph
}
