% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_subroutines.R
\name{fit_rw}
\alias{fit_rw}
\title{Pre-filter a track using a random walk model}
\usage{
fit_rw(locations, delta_t)
}
\arguments{
\item{locations}{An n x 2 sf data.frame (t, q, geom) giving the observed projected coordinates (point geometries), observation time, and location quality class of a movement path. The time column should be either a POSIXt column or a numeric column.}

\item{delta_t}{Time between locations in estimated true path, see ?seq.POSIXt}
}
\value{
A list
\itemize{
\item Locations A time-sorted copy of the passed in locations
\item Estimated An sf data.frame with columns
\itemize{
\item x_se, y_se Standard errors for the x and y coordinate, respectively
\item t Time for the location estimate
\item geom Point geometries giving the estimated locations
}
\item Parameters parameters estimates
}
}
\description{
Pre-filter a track using a random walk model
}
