% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nn_graph.R
\name{make_nn_graph}
\alias{make_nn_graph}
\title{Make a nearest neighbour grid optimized for a covariance function}
\usage{
make_nn_graph(xlim, ylim, cv_pars = c(1, 0.3, 2.5), cv_code = 1)
}
\arguments{
\item{xlim, }{ylim Vectors of length 2 giving the desired bounding box.}

\item{cv_pars}{Vector of length 3 - \link{marginal std. dev., range, smoothness}}

\item{cv_code}{0 = exponential, 1 = gaussian, 2 = matern, 3 = matern (nu = 1.5)}
}
\value{
A named list
\itemize{
\item stars: A stars object with the raster locations, variables, and values (w and se)
\item graph: A directed acyclic graph
}
}
\description{
Make a nearest neighbour grid optimized for a covariance function
}
