% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pred_graph.R
\name{make_pred_graph}
\alias{make_pred_graph}
\title{Make a nearest neighbour graph for a set of prediction locations}
\usage{
make_pred_graph(pred_coordinates, nn_graph)
}
\arguments{
\item{pred_coordinates}{The prediction locations as an sf data.frame with point geometries, with an additional column named "v" for the variable (1 = original, 2 = dx, 3 = dy).}

\item{nn_graph}{The output of \code{make_nn_graph}.}
}
\value{
A named list
\itemize{
\item v An integer vector specifying the field to be predicted.
\item coord A matrix of size n x 2 giving the prediction locations.
\item parents A list of integer matrices giving the index of locations used for the prediction.
}
}
\description{
Make a nearest neighbour graph for a set of prediction locations
}
