% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_starve_pred_graph.R
\name{make_starve_gg_pred_graph}
\alias{make_starve_gg_pred_graph}
\title{Create a prediction graph used to predict the (log) utilization distribution.}
\usage{
make_starve_gg_pred_graph(
  pred_coordinates,
  field_coordinates,
  cv_pars,
  cv_code = 1,
  k = 1
)
}
\arguments{
\item{pred_coordinates}{An sf data.frame with point geometries. The
locations for which predictions will be made.}

\item{field_coordinates}{An sf data.frame with point geometries. The
locations used in the mesh.}

\item{cv_pars}{Vector of length 2 or 3 - \link{std. dev., range, smoothness}}

\item{cv_code}{0 = exponential, 1 = gaussian, 2 = matern, 3 = matern (nu = 1.5)}

\item{k}{Number of parents in each direction}
}
\value{
A list with the prediction coordinates and list of parents.
}
\description{
Create a prediction graph used to predict the (log) utilization distribution.
}
