% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_starve_graph.R
\name{make_starve_graph}
\alias{make_starve_graph}
\title{Make a starve-style graph using an INLA mesh.}
\usage{
make_starve_graph(x, max.edge = 1, ...)
}
\arguments{
\item{x}{An sf object with point locations}

\item{max.edge}{The largest allowed triangle edge length. See INLA::inla.mesh.2d.}

\item{...}{Additional options to pass to INLA::inla.mesh.2d.}
}
\value{
A list with the mesh, node locations, and graph.
}
\description{
Make a starve-style graph using an INLA mesh.
}
