% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_starve_pred_graph.R
\name{make_starve_pred_graph}
\alias{make_starve_pred_graph}
\title{Create a prediction graph}
\usage{
make_starve_pred_graph(pred_coordinates, field_coordinates)
}
\arguments{
\item{pred_coordinates}{An sf data.frame with point geometries. The
locations for which predictions will be made.}

\item{field_coordinates}{An sf data.frame with point geometries. The
locations used in the mesh.}
}
\value{
A list with the prediction coordinates and list of parents.
}
\description{
Create a prediction graph
}
