% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_subroutines.R
\name{predict_utilization_distribution}
\alias{predict_utilization_distribution}
\title{Use a fitted langevin diffusion model to predict the utilization distribution}
\usage{
predict_utilization_distribution(
  prediction_locations,
  fitted_model,
  k = 1,
  ...
)
}
\arguments{
\item{prediction_locations}{An sf object with point geometries}

\item{fitted_model}{The output of fit_utilization_distribution}

\item{k}{The number of parents in each direction}
}
\value{
An sf object with predictions for the log-utilization distribution
}
\description{
Use a fitted langevin diffusion model to predict the utilization distribution
}
