% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_track}
\alias{simulate_track}
\title{Simulate a track from a Langevin diffusion model}
\usage{
simulate_track(
  xlim = c(-2, 2),
  ylim = c(-2, 2),
  boundary_sharpness = 2.5,
  boundary_limit = 0.8,
  cv_code = 1,
  cv_pars = c(2, 0.5, 6.5),
  nt = 100,
  nping = 40,
  gamma = 0.08,
  ping_tau = 0.1 * c(1, 1),
  ping_cor = 0.3,
  loc_class_probs = c(G = 0.026, `3` = 0.04, `2` = 0.035, `1` = 0.02, `0` = 0.065, A =
    0.126, B = 0.688),
  seed
)
}
\arguments{
\item{xlim, }{ylim The approximate bounding box for movement}

\item{boundary_sharpness}{How penalized should movement outside the bounding box be?}

\item{boundary_limit}{A scaling factor applied to xlim and ylim to determine where the penalty be applied.}

\item{cv_code}{0 = exponential (don't use), 1 = gaussian, 2 = matern, 3 = matern32}

\item{cv_pars}{c(marginal std. dev., range, smoothness)}

\item{nt}{How many true movement locations should there be?}

\item{nping}{How many location pings should there be?}

\item{gamma}{Speed parameter.}

\item{ping_tau}{Std. dev.s for ping observation error.}

\item{ping_cor}{Correlation between ping observation error coordinates.}

\item{loc_class_probs}{Named probability factor giving location quality class frequencies.}

\item{seed}{Optional simulation seed.}
}
\value{
A list:
\itemize{
\item true An sf data.frame with the true location track
\item pings An sf data.frame with the observation pings
}
}
\description{
Simulate a track from a Langevin diffusion model
}
