% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_subroutines.R
\name{fit_utilization_distribution}
\alias{fit_utilization_distribution}
\title{Fit a utilization distribution using a filtered movement track estimate}
\usage{
fit_utilization_distribution(
  track_estimate,
  cv_code = 1,
  initial_cv_pars = c(1, 1, 1.5),
  bbox = 1.1 * sf::st_bbox(track_estimate$pings)
)
}
\arguments{
\item{track_estimate}{The output of fit_rw or fit_langevin, e.g. a list
\itemize{
\item pings An sf data.frame with columns
\itemize{
\item t Time for the location ping
\item q Location quality class for the location ping
\item geom Point geometries giving the observed location pings
}
\item track An sf data.frame with columns
\itemize{
\item x_se, y_se Standard errors for the x and y coordinate, respectively
\item t Time for the location estimate
\item geom Point geometries giving the estimated locations
}
\item parameters parameter estimates
}}

\item{cv_code}{0 = exponential (don't use), 1 = gaussian, 2 = matern, 3 = matern32}

\item{bbox}{A named vector with elements "xmin", "xmax", "ymin", and "ymax"}

\item{The}{initial covariance function parameters used to build the nearest neighbour graph. Can either be a length 3 vector (std. dev., range, smoothness) or a k x 3 matrix. If a matrix is supplied, each row will be tested to find which gives the highest likelihood value.}
}
\description{
Fit a utilization distribution using a filtered movement track estimate
}
